Added more data types

+++ tests/Database/Adapters/DatabaseRelationalTest.php
@@ -72,7 +72,7 @@ class DatabaseRelationalTest extends TestCase
         $bookDateColumn = new Column('publication_date', ColumnType::TEXT);
         $bookDateColumn->setNotNull(false);
         $bookTable->addColumn($bookDateColumn);
-        $bookPriceColumn = new Column('price', ColumnType::REAL);
+        $bookPriceColumn = new Column('price', ColumnType::NUMERIC);
         $bookPriceColumn->setNotNull(true);
         $bookTable->addColumn($bookPriceColumn);
 
@@ -153,7 +153,7 @@ class DatabaseRelationalTest extends TestCase
         $passwordColumn = new Column('password', ColumnType::TEXT);
         $passwordColumn->setNotNull(true);
         $table->addColumn($passwordColumn);
-        $creditColumn = new Column('credit', ColumnType::REAL);
+        $creditColumn = new Column('credit', ColumnType::NUMERIC);
         $creditColumn->setNotNull(true);
         $table->addColumn($creditColumn);
         $registeredColumn = new Column('registered', ColumnType::DATETIME);
@@ -393,7 +393,7 @@ class DatabaseRelationalTest extends TestCase
         $table->addColumn($nameColumn);
         $authorColumn = new Column('author', ColumnType::TEXT);
         $table->addColumn($authorColumn);
-        $priceColumn = new Column('price', ColumnType::REAL);
+        $priceColumn = new Column('price', ColumnType::MONEY);
         $priceColumn->setNotNull(true);
         $table->addColumn($priceColumn);
 
@@ -487,7 +487,7 @@ class DatabaseRelationalTest extends TestCase
         $table->addColumn($nameColumn);
         $authorColumn = new Column('author', ColumnType::TEXT);
         $table->addColumn($authorColumn);
-        $priceColumn = new Column('price', ColumnType::REAL);
+        $priceColumn = new Column('price', ColumnType::NUMERIC);
         $priceColumn->setNotNull(true);
         $table->addColumn($priceColumn);
